metrics-core: switch entirely to Key for Recorder + new helper types
Recorder previously took a generic argument -- AsRef<str> -- to allow anything
that could fit that shape to act as a metric name.  We've now changed this to be
Into<Key>, which was previously called MetricName, which is a copy-on-write
string.

In practice, dynamic metric names are rare.  While certain systems/users
might choose to use an enum or other constantly-defined type when they
define their metrics on the write side, we ultimately deal with strings
internally, and when we push them into external systems.

This lets us push that to our consumers, which while not as maximally
flexible, trades call-site flexibility for high performance.  They can
still provide a custom implementation on top of their own enums or other
custom types, and work seamlessly with the Recorder trait, but they'll
be forced to do the work of generating the COW string for us.

We've also added a new type -- AsNanoseconds -- which allows us to generically
accept either a Duration or integer value when taking a value that should be
boiled down to a u64 nanosecond duration.  With this change, quanta will be
able to provide a wrapper type over u64 that enforces the duration being
nanoseconds, so that callers can require a known type that gives them
nanoseconds without worrying about a user unintentionally giving them
the wrong timebase.

Right now we have an implementation for u64 since quanta doesn't yet
have its own wrapper type, but that will go away in a future breaking
release.