feature: add a JSON observer
This adds a brand-new JSON observer to the mix.  We've refactored a bit,
making the builder for the observers actually function like a builder,
where methods to set the configuration update the builder fluently.
We've also extracted MetricsTree which is now serializable by serde.
This means anything based on MetricsTree can implicitly be serialized by
any serde serializer, which should make it easier for us to spin up new
output formats in the future.