Add new ways to snapshot the registry
The newly added methods provide more flexibility to the users, and they
build upon each other so users can only use the extra flexibility if they
want to.

We're also more careful with the Clone trait requirement for handles, as
for the most generic way the Clone is not required.