Switch from t1ha to aHash as the default hasher
t1ha causes a global-buffer-overflow when testing under ASAN. This
change switches the default hash implementation to aHash which has a
similar performance approach.

Switch to aHash removes the global-buffer-overflow and the tests
complete successfully.

In general, cargo bench shows performance improvements as high as 31% on
specific benches.

There are 3 regressions and 29 improvements total