Fix `metrics::Cow` provenance issue
Calling `.as_mut_ptr` on a `String` actually goes through `&mut str`,
which shrinks the provenance of the pointer to only contain the
initialized bytes. This caused issues when a reconstructed `String`
tried to write to the uninitialized part of it. The fix is to go through
`Vec::<u8>::as_mut_ptr`, which gives provenance for the entire
allocation.