Return self.min()/max() for the 0.0/1.0 quantiles
The 0.0 and 1.0 quantiles correspond to the minimum and maximum values
of the distribution. This change returns these values when the 0.0 and
1.0 quantiles are requested.

One benefit of doing so is that the minimum and maximum are the exact
values for these quantiles, rather than estimates. Another benefit is
that the 0.0 quantile can be considered a special case, since for any
quantile, we return the maximum value below that quantile. Since there
is no value below the 0.0 quantile, we instead return the minimum.