Bring the metrics-observer protobufs up to date
Aligned the proto file in the observer sub-crate with the one in
the TCP exporter. Updated the observer message handling code to use
the new structures. This format de-sync will occur again if the writer
code is changed without also updating the observer. The right solution
here is to have the two crates share the protobuf file, but that
requires more structural tinkering than a newcomer to the project
should do :).

Satisfied the clippy lints complaining about type complexity by
creeating some subtypes.