Clean up qualified type names.
The prost generated message wrapper type is named `Event`, and this
overlaps with our metrics Event type that is in its payload. Named
the wrapper `ProstMessage` to make its meaning more clear. This one
is only used once and might reasonably be used fully-qualified inline.

Also, fixed the typo in MetadataValue.