Add matrix setting to test MSRV distinct from Rust-toolchain
The rust-toolchain.toml config sets the version used to compile as 1.74,
while the MSRV for the rest of the workspace is 1.70. This is to allow
metrics-observer to build, as it requires 1.74.

This commit adds a matrix setting to the CI workflow to allow testing
the MSRV separately from the rest of the workspace, and some additional
documentation to explain the discrepancy.