dogstatsd: send length_prefix only for Unix Stream
Length prefix is only needed in case of Stream and not for Datagram whether
it Unix or UDP sockets.
Reference:
https://github.com/DataDog/datadog-go/blob/bae3560e5c664d64b71c0e8ca89326afa362e12a/statsd/uds.go#L62

Earlier, it was assumed that RemoteAddr::Unixgram has to also prefix length
which made the code complex to pad sent buffer with length, to avoid copy before sending.

Since length prefix is no longer needed, make the code simple by pushing the length padding
role until the end of send.