"Consume" counters and gauge on snapshot
This commit modifies the behavior of the snapshot to consume (or reset)
the observed counters and gauges. The motivation is to make them behave
in a similar way to histograms whose values are cleared from the backing
storage during snapshot. The counters and gauges are set to 0 as it is
their initial values at creation time.