histogram: change bucket behavior // comply with openmetrics
OpenMetrics [defines][openmetrics-histogram] histograms as cumulative.
A value should be assigned to a bucket and only one bucket, the sum of
the bucket is expected to be the number of values recorded:

```
Histogram MetricPoints MUST have one bucket with an +Inf threshold. Buckets MUST be cumulative. As an example for a metric representing request latency in seconds its values for buckets with thresholds 1, 2, 3, and +Inf MUST follow value_1 <= value_2 <= value_3 <= value_+Inf. If ten requests took 1 second each, the values of the 1, 2, 3, and +Inf buckets MUST equal 10.
```

This obviously changes the behavior of the histogram and is a breaking
change.

[openmetrics-histogram]: https://github.com/prometheus/OpenMetrics/blob/main/specification/OpenMetrics.md#histogram